requirements.txt
beautifulsoup4==4.5.1
html5lib==0.999999999
webencodings==0.5
boto3
appdirs==1.4.0
astroid==1.4.9
backports.functools-lru-cache==1.3
configparser==3.5.0
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.6.1
okta==0.0.3
packaging==16.8
pylint==1.6.5
pyparsing==2.1.10
PyYAML==3.12
regex==2017.1.17
requests==2.13.0
six==1.10.0
wrapt==1.10.8
pyldap==2.4.25.1
